requirements.txt
psycopg2-binary
SQLAlchemy==1.4.1
pandas==1.2.3
ipython-sql==0.4.0
python-dotenv==0.16.0